Renamed the Service APi config file, and moved the exception class into its own package
