+++ src/main/java/za/co/msrv/incubator/exceptions/ExternalAPIException.java
@@ -1,4 +1,4 @@
-package za.co.msrv.incubator.service;
+package za.co.msrv.incubator.exceptions;
 
 public class ExternalAPIException extends RuntimeException {
     public ExternalAPIException(String message) {
